project_files/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers>=4.40.0
torch>=2.6.0
supabase>=2.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
